Also change the types of `*_byte`

+++ src/frame/parser.rs
@@ -140,7 +140,7 @@ fn sample_or_frame_number(input: &[u8], is_sample: bool,
   }
 }
 
-fn secondary_block_size(input: &[u8], block_byte: u32)
+fn secondary_block_size(input: &[u8], block_byte: u8)
                         -> IResult<&[u8], Option<u32>> {
   match block_byte {
     0b0110 => opt!(input, map!(take!(1), to_u32)),
@@ -149,7 +149,7 @@ fn secondary_block_size(input: &[u8], block_byte: u32)
   }
 }
 
-fn secondary_sample_rate(input: &[u8], sample_byte: u32)
+fn secondary_sample_rate(input: &[u8], sample_byte: u8)
                         -> IResult<&[u8], Option<u32>> {
   match sample_byte {
     0b1100 => opt!(input, map!(take!(1), to_u32)),
